src/requirements.txt
pyemvue == 0.12.2
influx_line_protocol
paho-mqtt
